requirements.txt
gradio==4.7.1
openai>=1.40.0
mlflow==2.11.0
python-dotenv==1.0.0

